requirements.txt
chalice
requests
sqlalchemy
psycopg2-binary==2.9.5
aws_lambda_powertools==2.0.0
marshmallow
python-dotenv
pytz

